fix bugs of config
